refactor: 모든 프롬프트 생성은 PromptGenerator 에서 하도록 수정
